pyproject.toml
[project]
name = "tabby-config-sync"
version = "0.1.0"
description = "Tabby Config Sync"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "tortoise-orm[aiomysql,asyncpg]>=0.25.3",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
